requirements.txt
Flask-SQLAlchemy
flask
flask-migrate
Flask-WTF
jinja2
ruff